refactor(job-type-tracking): collapse user and onboarding endpoint onto user endpoint only + make jobtype optional
