+++ front/pages/api/user/index.ts
@@ -12,6 +12,8 @@ import logger from "@app/logger/logger";
 import { apiError } from "@app/logger/withlogging";
 import type { UserTypeWithWorkspaces, WithAPIErrorResponse } from "@app/types";
 import { sendUserOperationMessage } from "@app/types";
+import { isJobType } from "@app/types/job_type";
+import { renderLightWorkspaceType } from "@app/lib/workspace";
 
 export type PostUserMetadataResponseBody = {
   success: boolean;
@@ -20,6 +22,7 @@ export type PostUserMetadataResponseBody = {
 const PatchUserBodySchema = t.type({
   firstName: t.string,
   lastName: t.string,
+  jobType: t.union([t.string, t.undefined]),
 });
 
 export type GetUserResponseBody = {
@@ -80,10 +83,12 @@ async function handler(
           },
         });
       }
-      if (user.workspaces[0]?.role === "admin") {
+
+      const workspace = user.workspaces[0];
+      if (workspace?.role === "admin") {
         sendUserOperationMessage({
           message:
-            `workspace_sid: ${user.workspaces[0]?.sId}; email: [${user.email}]; ` +
+            `workspace_sid: ${workspace?.sId}; email: [${user.email}]; ` +
             `User Name [${user.firstName} ${user.lastName}].`,
           logger,
           channel: "C075LJ6PUFQ",
@@ -97,7 +102,9 @@ async function handler(
 
       const firstName = bodyValidation.right.firstName.trim();
       const lastName = bodyValidation.right.lastName.trim();
+      const jobType = bodyValidation.right.jobType?.trim();
 
+      // Update user's name
       if (firstName.length === 0 || lastName.length === 0) {
         return apiError(req, res, {
           status_code: 400,
@@ -109,6 +116,37 @@ async function handler(
       }
 
       await u.updateName(firstName, lastName);
+
+      // Update user's jobType
+      if (jobType !== undefined && !isJobType(jobType)) {
+        return apiError(req, res, {
+          status_code: 400,
+          api_error: {
+            type: "invalid_request_error",
+            message: "Job type is invalid.",
+          },
+        });
+      }
+
+      // metadata + for loop allows for
+      // more metadata to be processed thru
+      // endpoint in future
+      const metadata = {
+        job_type: jobType,
+      };
+
+      for (const [key, value] of Object.entries(metadata)) {
+        if (value !== undefined) {
+          await u.setMetadata(key, String(value));
+        }
+      }
+
+      await ServerSideTracking.trackCompleteUserOnboarding({
+        user: user,
+        workspace: renderLightWorkspaceType({ workspace }),
+        role: workspace.role !== "none" ? workspace.role : "user",
+        jobType,
+      });
 
       res.status(200).json({
         success: true,
